update type data

+++ lib/helper/currency_helper.dart
@@ -1,7 +1,7 @@
 import 'package:intl/intl.dart';
 
 class CurrencyHelper {
-  static String format(int value) {
+  static String format(double value) {
     final formatCurrency = NumberFormat.simpleCurrency(locale: 'id_ID');
     return formatCurrency.format(value);
   }
